finish update dependencies
